Preparing for visual overhaul.
