+++ Honehoover/Honehoover/Honehoover.vcxproj.filters
@@ -41,6 +41,9 @@
     <ClCompile Include="..\src\game_scores.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="..\src\game_getter.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\hdr\board.h">
