+++ Honehoover/Honehoover/highscores.txt
@@ -1,6 +1,7 @@
+13 Easy
 19 Easy
+21 Easy
 114 Medium
-13 Custom
 67 Easy
 103 Easy
 108 Easy
@@ -12,4 +13,3 @@
 508 Easy
 513 Easy
 603 Easy
-608 Easy
