+++ Honehoover/src/game_master.cpp
@@ -77,98 +77,12 @@ void Game::run() {
     }
 }
 
-const Game::Difficulty Game::difficulties[] = {
-    { 12, 9, 4 },
-    { 20, 15, 20 }, 
-    { 30, 24, 184 }
-};
-
-void Game::setDifficulty(DifficultyLevel level) {
-    switch (level) {
-    case DifficultyLevel::EASY:
-        board = Board(12, 9, 4);
-        break;
-    case DifficultyLevel::MEDIUM:
-        board = Board(20, 15, 20);
-        break;
-    case DifficultyLevel::HARD:
-        board = Board(30, 24, 184);
-        break;
-    case DifficultyLevel::CUSTOM:
-            board = Board(customDifficulty.width, customDifficulty.height, customDifficulty.mineCount);
-            if (customDifficulty == difficulties[0]) { //0
-                currentDifficulty = DifficultyLevel::EASY;
-            }
-            else if (customDifficulty == difficulties[1]) { //32
-                currentDifficulty = DifficultyLevel::MEDIUM;
-            }
-            else if (customDifficulty == difficulties[2]) { //72
-                currentDifficulty = DifficultyLevel::HARD;
-            }
-            else {
-                currentDifficulty = DifficultyLevel::CUSTOM;
-            }
-            break;
-    }
-    currentDifficulty = level;
-}
-
-void Game::setCustomDifficulty(int sliderVal)
-{
-    int width = 12 + sliderVal * 5 / 20;
-    int height = 9 + sliderVal * 5 / 24;
-    int mineCount = 4 + sliderVal * (0.5 + (sliderVal * sliderVal) / 2000);
-    customDifficulty = { width, height, mineCount, true };
-}
-
-string Game::difficultyToString(DifficultyLevel level) {
-    switch (level) {
-    case DifficultyLevel::EASY:
-        return "Easy";
-    case DifficultyLevel::MEDIUM:
-        return "Medium";
-    case DifficultyLevel::HARD:
-        return "Hard";
-    case DifficultyLevel::CUSTOM:
-        return "Custom";
-    default:
-        return "Unknown Difficulty";
-    }
-}
-
-int Game::getDifficultySliderValue(DifficultyLevel level) {
-    switch (level) {
-    case DifficultyLevel::EASY:
-        return 0; 
-    case DifficultyLevel::MEDIUM:
-        return 32;
-    case DifficultyLevel::HARD:
-        return 72; 
-    default:
-        return -1;
-    }
-}
-
-Game::DifficultyLevel Game::getCurrentDifficulty(int sliderVal) {
-    if (abs(sliderVal - getDifficultySliderValue(DifficultyLevel::EASY)) < 5) {
-        return DifficultyLevel::EASY;
-    }
-    else if (abs(sliderVal - getDifficultySliderValue(DifficultyLevel::MEDIUM)) < 5) {
-        return DifficultyLevel::MEDIUM;
-    }
-    else if (abs(sliderVal - getDifficultySliderValue(DifficultyLevel::HARD)) < 5){
-        return DifficultyLevel::HARD;
-    }
-    return DifficultyLevel::CUSTOM;
-}
-
 template<typename T>
 const T& clamp(const T& val, const T& lower, const T& upper) {
     if (val < lower) return lower;
     else if (val > upper) return upper;
     else return val;
 }
-
 
 void Game::handleEvents() {
     SDL_Event e;
@@ -285,6 +199,46 @@ void Game::update() {
         // Update flag count
         remainingFlags = board.getMineCount() - board.getFlagCount();
     }
+}
+
+
+
+void Game::setDifficulty(DifficultyLevel level) {
+    switch (level) {
+    case DifficultyLevel::EASY:
+        board = Board(12, 9, 4);
+        break;
+    case DifficultyLevel::MEDIUM:
+        board = Board(20, 15, 20);
+        break;
+    case DifficultyLevel::HARD:
+        board = Board(30, 24, 184);
+        break;
+    case DifficultyLevel::CUSTOM:
+        board = Board(customDifficulty.width, customDifficulty.height, customDifficulty.mineCount);
+        if (customDifficulty == difficulties[0]) { //0
+            currentDifficulty = DifficultyLevel::EASY;
+        }
+        else if (customDifficulty == difficulties[1]) { //32
+            currentDifficulty = DifficultyLevel::MEDIUM;
+        }
+        else if (customDifficulty == difficulties[2]) { //72
+            currentDifficulty = DifficultyLevel::HARD;
+        }
+        else {
+            currentDifficulty = DifficultyLevel::CUSTOM;
+        }
+        break;
+    }
+    currentDifficulty = level;
+}
+
+void Game::setCustomDifficulty(int sliderVal)
+{
+    int width = 12 + sliderVal * 5 / 20;
+    int height = 9 + sliderVal * 5 / 24;
+    int mineCount = 4 + sliderVal * (0.5 + (static_cast<double>(sliderVal) * sliderVal) / 2000);
+    customDifficulty = { width, height, mineCount, true };
 }
 
 void Game::resetGame() {
